[XrdCrypto][XrdHttp] Extract DN from user (proxy, multi-proxy) certificate and properly   handle the gridmap-file functionality when accessing through HTTP
