+++ src/XrdCrypto/XrdCryptoFactory.hh
@@ -55,6 +55,7 @@ class XrdCryptoX509Chain;
 class XrdCryptogsiX509Chain;
 class XrdCryptoX509Crl;
 class XrdCryptoX509Req;
+class XrdTlsPeerCerts;
 
 //
 // Prototypes for some Utility Functions
@@ -79,6 +80,11 @@ typedef int (*XrdCryptoX509ChainToFile_t)(XrdCryptoX509Chain *, const char *);
 // certificates from file parsing
 typedef int (*XrdCryptoX509ParseFile_t)(const char *fname,
                                         XrdCryptoX509Chain *);
+
+// certificates from STACK_OF(X509*)
+typedef int (*XrdCryptoX509ParseStack_t)(XrdTlsPeerCerts* pc,
+                                         XrdCryptoX509Chain *c);
+
 // certificates from bucket parsing
 typedef int (*XrdCryptoX509ParseBucket_t)(XrdSutBucket *,
                                           XrdCryptoX509Chain *);
@@ -173,6 +179,7 @@ public:
    virtual XrdCryptoX509VerifyCert_t X509VerifyCert();
    virtual XrdCryptoX509VerifyChain_t X509VerifyChain();
    virtual XrdCryptoX509ParseFile_t X509ParseFile();
+   virtual XrdCryptoX509ParseStack_t X509ParseStack();
    virtual XrdCryptoX509ParseBucket_t X509ParseBucket();
    virtual XrdCryptoX509ExportChain_t X509ExportChain();
    virtual XrdCryptoX509ChainToFile_t X509ChainToFile();
