+++ src/XrdCrypto/XrdCryptosslAux.hh
@@ -41,6 +41,9 @@
 
 #define kSslKDFunDefLen  24
 
+//! Froward declaration
+class XrdTlsPeerCerts;
+
 //
 // Password-Based Key Derivation Function 2, specified in PKCS #5
 //
@@ -60,6 +63,8 @@ int XrdCryptosslX509ChainToFile(XrdCryptoX509Chain *c, const char *fn);
 int XrdCryptosslX509ParseFile(const char *fname, XrdCryptoX509Chain *c);
 // certificates from bucket parsing
 int XrdCryptosslX509ParseBucket(XrdSutBucket *b, XrdCryptoX509Chain *c);
+// certificates from STACK_OF(X509*)
+int XrdCryptosslX509ParseStack(XrdTlsPeerCerts* pc, XrdCryptoX509Chain *chain);
 //
 // Function to convert from ASN1 time format into UTC since Epoch (Jan 1, 1970) 
 time_t XrdCryptosslASN1toUTC(const ASN1_TIME *tsn1);
